Move fee stuff to mod
